chore(wdio): port key-reorder test suite to wdio
