--- test/karma/test-app/key-reorder/index.html
@@ -1,6 +1,0 @@
-<!DOCTYPE html>
-<meta charset="utf8">
-<script src="./build/testapp.esm.js" type="module"></script>
-<script src="./build/testapp.js" nomodule></script>
-
-<key-reorder-root></key-reorder-root>
